requirements.txt
pandas
xlrd
xlwt
requests
pillow
pytesseract
selenium